Add files via upload
A function that takes in a string containing the name of a location in the form ('city, state, zip, country'), a start date as either a string in the form 'YYYY-MM-DD' or a datetime object, and an optional end date of the same form. Note not all of the fields listed for place need to be included and if no end date is given the default is the current date.